requirements.txt
numpy
scipy
opencv-python
scikit-image
typer[all]>=0.9.0
rich>=13.0.0
tifffile>=2023.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
huggingface-hub>=0.16.0
Pillow>=9.0.0
timm>=0.9.0
matplotlib>=3.5.0
pyparsing>=3.0.0

# BioMedCLIP dependencies 
open_clip_torch>=2.24.0
transformers>=4.40.0
